minor clarification about deriving Copy and Clone
Please feel free to reject this, but I found this sentence to be very confusing. I didn't understand the connection between the first sentence ("You can derive Copy on any type whose ...") with the next ("You can only apply the Copy trait ..."). I had to read the reference, and found its explanation to be clearer:

https://doc.rust-lang.org/std/marker/trait.Copy.html

"Clone is a supertrait of Copy, so everything which is Copy must also implement Clone. If a type is Copy then its Clone implementation only needs to return *self (see the example above)."

I'm not sure we need to get into the supertrait part, but I tried to make the explanation here similar to the reference versions.